[SofaFramework] Relocate functional tests in *_simutest
Those tests were in frameworkextra_test before.
