+++ SofaKernel/modules/SofaDefaultType/SofaDefaultType_simutest/CMakeLists.txt
@@ -1,0 +1,14 @@
+cmake_minimum_required(VERSION 3.1)
+
+project(SofaDefaultType_simutest)
+
+find_package(SofaComponentBase REQUIRED)
+
+set(SOURCE_FILES
+    TemplateAliases_test.cpp
+)
+
+add_executable(${PROJECT_NAME} ${SOURCE_FILES})
+target_link_libraries(${PROJECT_NAME} SofaGTestMain SofaComponentBase)
+
+add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME})
