+++ SofaKernel/modules/SofaDefaultType/SofaDefaultType_test/CMakeLists.txt
@@ -5,11 +5,10 @@ project(SofaDefaultType_test)
 set(SOURCE_FILES
     MapMapSparseMatrixEigenUtils_test.cpp
     MatTypes_test.cpp
-    TemplateAliases_test.cpp
     VecTypes_test.cpp
 )
 
 add_executable(${PROJECT_NAME} ${SOURCE_FILES})
-target_link_libraries(${PROJECT_NAME} SofaGTestMain SofaHelper)
+target_link_libraries(${PROJECT_NAME} SofaGTestMain SofaDefaultType)
 
 add_test(NAME SofaDefaultType_test COMMAND SofaDefaultType_test)
